+++ lib/local/CLM/include/SVR_patch_expert.h
@@ -46,12 +46,8 @@
 //       in IEEE Int. Conference on Computer Vision Workshops, 300 Faces in-the-Wild Challenge, 2013.    
 //
 ///////////////////////////////////////////////////////////////////////////////
-
 #ifndef __SVR_PATCH_EXPERT_h_
 #define __SVR_PATCH_EXPERT_h_
-
-#include <map>
-#include <cv.h>
 
 using namespace cv;
 
@@ -104,8 +100,8 @@ class SVR_patch_expert{
 		void Read(std::ifstream &stream);
 
 		// The actual response computation from intensity or depth (for CLM-Z)
-		void Response(const Mat_<float> &area_of_interest, Mat_<double> &response);    
-		void ResponseDepth(const Mat_<float> &area_of_interest, Mat_<double> &response);
+		void Response(const Mat_<float> &area_of_interest, Mat_<float> &response);    
+		void ResponseDepth(const Mat_<float> &area_of_interest, Mat_<float> &response);
 
 };
 //===========================================================================
@@ -135,8 +131,8 @@ class Multi_SVR_patch_expert{
 		void Read(std::ifstream &stream);
 
 		// actual response computation from intensity of depth (for CLM-Z)
-		void Response(const Mat_<float> &area_of_interest, Mat_<double> &response);
-		void ResponseDepth(const Mat_<float> &area_of_interest, Mat_<double> &response);
+		void Response(const Mat_<float> &area_of_interest, Mat_<float> &response);
+		void ResponseDepth(const Mat_<float> &area_of_interest, Mat_<float> &response);
 
 };
 }
